new DB admin - fav tour - get, put, post, delete (backend e frontend)

+++ BackEndFinalProject/src/db/tourGuideDB.js
@@ -15,7 +15,6 @@ async function getTourGuideFromDatabase() {
 	guides.guide_id = tours_guides.guide_id;`;
 
 	const response = await connection.promise().query(sql);
-
 	const result = response[0];
 	return result;
 }
@@ -51,9 +50,7 @@ async function getTourGuideByTourIDFromDatabase(id) {
 	ON
 	guides.guide_id = tours_guides.guide_id
 	WHERE
-	tours.tour_id = ?
-	GROUP BY
-	tours.tour_name`;
+	tours.tour_id = ?`;
 
 	const params = [id];
 
@@ -66,28 +63,38 @@ async function getTourGuideByTourIDFromDatabase(id) {
 async function insertNewTourGuideToDatabase(tourGuide) {
 	const sql = 'INSERT INTO tours_guides VALUES (?, ?) ';
 	const params = [tourGuide.tour_id, tourGuide.guide_id];
+	try {
+		const result = await connection.promise().query(sql, params);
 
-	const response = await connection.promise().query(sql, params);
-
-	return response;
+		return result;
+	} catch (error) {
+		throw error;
+	}
 }
 
 async function updateTourGuideFromDatabase(newGuide_id, [tour_id, guide_id]) {
-	const sql = 'UPDATE tours_guides SET guide_id = ? WHERE tour_id = ? AND guide_id = ?';
-
-	const params = [newGuide_id, tour_id, guide_id];
-	console.log('params', params);
-	const response = await connection.promise().query(sql, params);
-	console.log('response', response);
-	return response;
+	if (guide_id != 4) {
+		const sql = 'UPDATE tours_guides SET guide_id = ? WHERE tour_id = ? AND guide_id = ?';
+		const params = [newGuide_id, tour_id, guide_id];
+		console.log('params', params);
+		const response = await connection.promise().query(sql, params);
+		console.log('response', response);
+		return response;
+	} else {
+		throw new Error('You cannot update guide_id = 4!!!');
+	}
 }
 
 async function deleteTourGuideFromDatabase(guide_id, tour_id) {
-	const sql = 'DELETE FROM tours_guides WHERE guide_id = ? AND tour_id = ?';
+	if (guide_id != 4) {
+		const sql = 'DELETE FROM tours_guides WHERE guide_id = ? AND tour_id = ?';
 
-	const response = await connection.promise().query(sql, [guide_id, tour_id]);
+		const response = await connection.promise().query(sql, [guide_id, tour_id]);
 
-	return response;
+		return response;
+	} else {
+		throw new Error('You cannot delete guide_id = 4!!!');
+	}
 }
 
 module.exports = {
